Prevent logging npm based error message after lint
If eslint finds some issues then it will return a non-zero exit code
and so npm thinks that there was an error, but this isn't true.
While it is correct behaviour of eslint to return an error code as
exit status if there were found some issues the script itself did its
job very well.
Therefore even if the exit code is non-zero npm will get a zero code to
prevent the aftercomming error messages.